+++ src/arena.rs
@@ -2,12 +2,9 @@ use std::convert::TryInto;
 use std::mem::replace;
 use std::ops;
 
-use crate::drain::Drain;
 use crate::free_pointer::FreePointer;
 use crate::generation::Generation;
-use crate::into_iter::IntoIter;
-use crate::iter::Iter;
-use crate::iter_mut::IterMut;
+use crate::iter::{Drain, IntoIter, Iter, IterMut};
 
 /// Container that can have elements inserted into it and removed from it.
 ///