+++ src/iter/mod.rs
@@ -1,0 +1,11 @@
+//! Contains all of the iterator types for Thunderdome.
+
+mod drain;
+mod into_iter;
+mod iter;
+mod iter_mut;
+
+pub use drain::Drain;
+pub use into_iter::IntoIter;
+pub use iter::Iter;
+pub use iter_mut::IterMut;